Support TLS in the worker service endpoint
We only need this for the worker service as the rest of the endpoints are under an ingress controller that terminates TLS
before requests hit the endpoints.

The worker service endpoint is different from the rest of the endpoints as they can be directly hit by clients (as
an ingress controller might not support gRPC streaming).