Avoid a deadlock when a context is cannceled
When the context is canceled from WaitForCompletion,
we no longer read from the channels of the task, but the
processor still attempts to write when it receives a message from the engine.

We need to spawn a gorountine to read messages whenever comes. We could
also increase the channel buffer to 1 (and keep sufficient buffer for resp.Body), but
just reading from channels seems simpler.