Decopule puller from Ollama
With this change, 'puller' will simply downloads the model and create a modelfile, but it doesn't actually
register the model in Ollama.

We create another init container that runs 'ollama serve' and 'ollama create'.

This will also help us support additional runtimes in the future as the puller code can be simply reused.