Refactor the HuggingFace download logic to explicitly take an adapter type as argument
Some of the existing setup doesn't have the ListBucket permission. Removing the list call
from the default path is required to avoid regression.

Also it would be better if we can explictily check if all necessary files exist or not
depending on whether a requested model is adapter or not.